+++ UIImage+PDF/UIImage+PDF.h
@@ -23,50 +23,50 @@
 +(UIImage *) imageOrPDFWithContentsOfFile:(NSString *)path;
 
 
-+(UIImage *) imageWithPDFNamed:(NSString *)resourceName atSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFNamed:(NSString *)resourceName atSize:(CGSize)size atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFNamed:(NSString *)resourceName atSize:(CGSize)size;
 
-+(UIImage *) imageWithPDFNamed:(NSString *)resourceName atWidth:(CGFloat)width atPage:(int)page;
++(UIImage *) imageWithPDFNamed:(NSString *)resourceName atWidth:(CGFloat)width atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFNamed:(NSString *)resourceName atWidth:(CGFloat)width;
 
-+(UIImage *) imageWithPDFNamed:(NSString *)resourceName atHeight:(CGFloat)height atPage:(int)page;
++(UIImage *) imageWithPDFNamed:(NSString *)resourceName atHeight:(CGFloat)height atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFNamed:(NSString *)resourceName atHeight:(CGFloat)height;
 
-+(UIImage *) imageWithPDFNamed:(NSString *)resourceName fitSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFNamed:(NSString *)resourceName fitSize:(CGSize)size atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFNamed:(NSString *)resourceName fitSize:(CGSize)size;
 
-+(UIImage *) originalSizeImageWithPDFNamed:(NSString *)resourceName atPage:(int)page;
++(UIImage *) originalSizeImageWithPDFNamed:(NSString *)resourceName atPage:(NSUInteger)page;
 +(UIImage *) originalSizeImageWithPDFNamed:(NSString *)resourceName;
 
 
-+(UIImage *) imageWithPDFURL:(NSURL *)URL atSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFURL:(NSURL *)URL atSize:(CGSize)size atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFURL:(NSURL *)URL atSize:(CGSize)size;
 
-+(UIImage *) imageWithPDFURL:(NSURL *)URL atWidth:(CGFloat)width atPage:(int)page;
++(UIImage *) imageWithPDFURL:(NSURL *)URL atWidth:(CGFloat)width atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFURL:(NSURL *)URL atWidth:(CGFloat)width;
 
-+(UIImage *) imageWithPDFURL:(NSURL *)URL atHeight:(CGFloat)height atPage:(int)page;
++(UIImage *) imageWithPDFURL:(NSURL *)URL atHeight:(CGFloat)height atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFURL:(NSURL *)URL atHeight:(CGFloat)height;
 
-+(UIImage *) imageWithPDFURL:(NSURL *)URL fitSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFURL:(NSURL *)URL fitSize:(CGSize)size atPage:(NSUInteger)page;
 +(UIImage *) imageWithPDFURL:(NSURL *)URL fitSize:(CGSize)size;
 
 
 +(UIImage *) imageWithPDFData:(NSData *)data atSize:(CGSize)size;
-+(UIImage *) imageWithPDFData:(NSData *)data atSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFData:(NSData *)data atSize:(CGSize)size atPage:(NSUInteger)page;
 
 +(UIImage *) imageWithPDFData:(NSData *)data atWidth:(CGFloat)width;
-+(UIImage *) imageWithPDFData:(NSData *)data atWidth:(CGFloat)width atPage:(int)page;
++(UIImage *) imageWithPDFData:(NSData *)data atWidth:(CGFloat)width atPage:(NSUInteger)page;
 
 +(UIImage *) imageWithPDFData:(NSData *)data atHeight:(CGFloat)height;
-+(UIImage *) imageWithPDFData:(NSData *)data atHeight:(CGFloat)height atPage:(int)page;
++(UIImage *) imageWithPDFData:(NSData *)data atHeight:(CGFloat)height atPage:(NSUInteger)page;
 
 +(UIImage *) imageWithPDFData:(NSData *)data fitSize:(CGSize)size;
-+(UIImage *) imageWithPDFData:(NSData *)data fitSize:(CGSize)size atPage:(int)page;
++(UIImage *) imageWithPDFData:(NSData *)data fitSize:(CGSize)size atPage:(NSUInteger)page;
 
 
 +(UIImage *) originalSizeImageWithPDFData:(NSData *)data;
-+(UIImage *) originalSizeImageWithPDFURL:(NSURL *)URL atPage:(int)page;
++(UIImage *) originalSizeImageWithPDFURL:(NSURL *)URL atPage:(NSUInteger)page;
 +(UIImage *) originalSizeImageWithPDFURL:(NSURL *)URL;
